BUGFIX: Don't send value for replaced assets
previously when no asset was selected or a previously selected asset
was replaced with the `asset-replace` element, an empty string (`""`)
was submitted instead of the asset identifier.

Now the field element is disabled in that case leading to no value (= `null`)
being sent for a removed asset.

This might improve compatibility with the Flow PropertyMapper that can map
`null` to a property of type `AssetInterface|null`.

Fixes: #3